Unify plain translation handling

+++ translations/handlePlainTranslations.sh
@@ -9,10 +9,12 @@ gpg --allow-secret-key-import --import /gpg/nextcloud-bot.asc
 gpg --list-keys
 
 # fetch git repo
-git clone git@github.com:nextcloud/android /app
+git clone git@github.com:$1/$2 /app
 
 # remove existing translations to cleanup not maintained languages
-rm -rf src/main/res/values-*/strings.xml
+if [ -d src/main/res ]; then
+  rm -rf src/main/res/values-*/strings.xml
+fi
 
 # push sources
 tx push -s
@@ -20,8 +22,10 @@ tx push -s
 # pull translations
 tx pull -f -a --minimum-perc=50
 
-# copy transifex strings to fastlane
-python3 scripts/metadata/generate_metadata.py
+if [ -e "scripts/metadata/generate_metadata.py" ]; then
+  # copy transifex strings to fastlane
+  python3 scripts/metadata/generate_metadata.py
+fi
 
 # create git commit and push it
 git add .
